Improved query performance and data access design, refactored lat/long

+++ DataSetExtension.Tests/FakeMeasurementLocator.cs
@@ -14,7 +14,7 @@ namespace DataSetExtension.Tests
 		
 		public StationTracker Tracker { get; set; }
 		
-		public Measurement Find(decimal latitude, decimal longitude, DateTime date) 
+		public Measurement Find(double latitude, double longitude, DateTime date) 
 		{
 			IsNew = true;
