+++ DataSetExtension/Boundry.cs
@@ -6,12 +6,12 @@ namespace DataSetExtension
 {
     public class Boundry
     {
-        public decimal MinLatitude { get; set; }
+        public double MinLatitude { get; set; }
  
-        public decimal MaxLatitude { get; set; }
+        public double MaxLatitude { get; set; }
  
-        public decimal MinLongitude { get; set; }
+        public double MinLongitude { get; set; }
  
-        public decimal MaxLongitude { get; set; }
+        public double MaxLongitude { get; set; }
     }
 }
